\name{hello}
\alias{hello}
\title{Hello, World!}
\usage{
hello()
}
\description{
Prints 'Hello, world!'.
}
\examples{
hello()
}

\name{small_cl}
\alias{small_cl}
\title{This function paste a random cell ramdomly in reigbor in a cluster}
\usage{
small_cl()
}
\description{
small_cl 'Hello, world!'.
}
\examples{
data<-clusters_matrix(5,.5)
small_cl(data)
}
